requirements.txt
Django~=5.0.1
djangorestframework~=3.14.0
channels~=4.0.0
asgiref~=3.7.2
